+++ src/main/java/com/daw2/proyectoFinal/servicesImpl/ProyectoServiceImpl.java
@@ -1,5 +1,7 @@
 package com.daw2.proyectoFinal.servicesImpl;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
@@ -7,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.daw2.proyectoFinal.model.Proyecto;
+import com.daw2.proyectoFinal.model.Tarea;
 import com.daw2.proyectoFinal.repository.ProyectoRepository;
 import com.daw2.proyectoFinal.services.ProyectoService;
 
@@ -33,8 +36,22 @@ public class ProyectoServiceImpl implements ProyectoService {
     }
 
     @Override
-    public void eliminarProyecto(Long id) {
-        proyectoRepository.deleteById(id);
+    public boolean eliminarProyecto(Long id) {
+        if (proyectoRepository.existsById(id)) {
+        	proyectoRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+
+	@Override
+	public Proyecto actualizarProyecto(Proyecto proyecto) {
+		return proyectoRepository.save(proyecto);
+	}
+    
+	@Override
+    public List<Proyecto> obtenerProyectosConUsuario(Long usuarioId) {
+        return proyectoRepository.findByUsuariosId(usuarioId);
     }
 
 }
