+++ src/main/java/com/daw2/proyectoFinal/servicesImpl/UsuarioServiceImpl.java
@@ -1,12 +1,17 @@
 package com.daw2.proyectoFinal.servicesImpl;
 
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.daw2.proyectoFinal.model.Proyecto;
+import com.daw2.proyectoFinal.model.Tarea;
 import com.daw2.proyectoFinal.model.Usuario;
+import com.daw2.proyectoFinal.repository.ProyectoRepository;
 import com.daw2.proyectoFinal.repository.UsuarioRepository;
 import com.daw2.proyectoFinal.services.UsuarioService;
 
@@ -15,6 +20,9 @@ public class UsuarioServiceImpl implements UsuarioService {
 
     @Autowired
     private UsuarioRepository usuarioRepository;
+    
+    @Autowired
+    private ProyectoRepository proyectoRepository;
 
     @Override
     public Usuario crearUsuario(Usuario usuario) {
@@ -33,8 +41,26 @@ public class UsuarioServiceImpl implements UsuarioService {
     }
 
     @Override
-    public void eliminarUsuario(Long id) {
-        usuarioRepository.deleteById(id);
+    public boolean eliminarUsuario(Long id) {
+        if (usuarioRepository.existsById(id)) {
+            usuarioRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
+    @Override
+    public Usuario obtenerUsuarioPorNombre(String nombreUsuario) {
+        return usuarioRepository.findByNombreUsuario(nombreUsuario);
+    }
+    
+    @Override
+    public Usuario actualizarUsuario(Usuario usuario) {
+        return usuarioRepository.save(usuario);
+    }
+    
+    @Override
+    public List<Proyecto> obtenerProyectosDeUsuario(Long usuarioId) {
+        return proyectoRepository.findByUsuariosId(usuarioId);
+    }
 }
